Pull-up empty methods from login and register activities to basic
In order to clean up child-class of basic.
